refactor(common/Table): restructure & restyle
